+++ src/main/java/io/github/byzatic/utility/prometheus/exporter/docker_raid_metrics_prometheus_exporter/collector/smartctl/dto/read/SmartctlDiskJson.java
@@ -34,6 +34,6 @@ public class SmartctlDiskJson {
     }
 
     public static class RawValue {
-        public int value;
+        public long value;
     }
 }
